New HitBoxes Implemented
Hitboxes now rotate with the object and surround it, making sure that all collisions are detection.

The added problem now is that the Physics systems need to be properly implemented, as some behaviour is weird.

i.e. - Simply taking the velocity of the robot and added it to an object, if you've droven past the object and are hitting on the turn, it shouldn't "pick it up"